move tests to test_getitem.py
